refactor creating app
