+++ auth/src/index.ts
@@ -1,41 +1,9 @@
-import express from 'express';
-import 'express-async-errors'; // we need to prevent error in sync functions
-import { json } from 'body-parser';
 import mongoose from 'mongoose';
-import cookieSession from 'cookie-session'
-
-import { currentUserRouter } from './routes/current-user';
-import { signinRouter } from './routes/signin';
-import { signoutRouter } from './routes/signout';
-import { signupRouter } from './routes/signup';
-import { errorHandler } from './middlewares/error-handler';
-import { NotFoundError } from './errors/not-found-error';
-
-const app = express();
-app.set('trust proxy', true);
-app.use(json());
-
-app.use(
-  cookieSession({
-    signed: false,
-    secure: true,
-  })
-)
-
-app.use(currentUserRouter);
-app.use(signinRouter);
-app.use(signoutRouter);
-app.use(signupRouter);
-
-app.all('*', async (req, res) => {
-  throw new NotFoundError();
-});
-
-app.use(errorHandler);
+import { app } from './app';
 
 const start = async () => {
 
-  if(!process.env.JWT_KEY) {
+  if (!process.env.JWT_KEY) {
     throw new Error('JWT_KEY must be defined');
   }
